index.ts and dexie-static-props complete. Now, what's left is the implementation of the Dexie class methods + fix typings tests + build script.
